Get cantonal ballot dates

Get cantonal ballot dates

Usage

ballot_dates_cantonal(
  cantons = c("Geneva", "Neuchatel", "Zurich"),
  exclude_counterproposals = TRUE,
  use_cache = TRUE,
  max_cache_age = "1 day"
)

Arguments

cantons

The names of the cantons to retrieve ballot dates for. A character vector.

exclude_counterproposals

Exclude indirect counterproposals and tie-breaker questions. A logical scalar.

use_cache

Whether or not to return cached results if possible. If FALSE, results are always newly fetched regardless of max_cache_age.

max_cache_age

Duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Use Inf to disable cache expiry. Only relevant if use_cache = TRUE.

Value

pkgsnippets::return_label("data")