Read in raw data files from the canton of Geneva

Read in raw data files from the canton of Geneva

Usage

read_raw_data_geneva(
  ballot_date,
  specific_datasets = NULL,
  check_municipalities = TRUE,
  remove_duplicated_votes = TRUE,
  remove_unknown_municipality_district_combos = FALSE,
  remove_na_locality_codes = FALSE,
  skip_problematic_dates = FALSE
)

Arguments

ballot_date

The ballot date of which to read in raw data files.

specific_datasets

If not NULL, only the specified filenames will be processed. A list with ballot dates as keys and character vectors of specific filenames (without the filetype extension) as values.

check_municipalities

Check if municipality could be determined for all entries excluding Swiss living abroad. A logical scalar.

remove_duplicated_votes

Remove duplicate votes (where votcorr == "A" | votdup == "1"; only if the necessary variable exists). A logical scalar. The variable votcorr is used to identify voters for whom several voting cards have been issued (due to non-receipt, loss, etc.). The dummy variable votdup serves the exact same purpose. The cases with votcorr = "A" (or votdup = "1") have to be filtered out in order to get correct results. In the (very rare) case both variables are included and contradict each other, votdup is given preference.

remove_unknown_municipality_district_combos

Ignore rows in raw data that are not Swiss abroad but missing a municipality (usually with the undocumented district code 99). A logical scalar.

remove_na_locality_codes

Ignore rows in raw data that have no municipality_code_2_digits set (NA). A logical scalar.

skip_problematic_dates

Skip ballot_dates for which unresolved parsing problems exist. A logical scalar.

Value

A list with a tibble of each raw data file plus metadata tibble.